consumer/requirements.txt
kafka-python
fastapi
uvicorn
numpy
opencv-python
urllib3
prometheus-client
